Add new org api parameters.
Type and Affiliation/Visibility are exclusive, if one is set the other
can't be, so the enum properties have to be nullable. The default
behaviour when no parameters are passed is whatever the new default
behaviour is on the server for the new api (currently, it is
Visibility = All and Affiliation = Owner | Collaborator.